+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # unreleased
+* Fallback to lookup tokens from other prints when card has no tokens but rules text mentions tokens
+
+# 0.12.0
 * Add `post` method (closes #9)
 * Add `getSymbolUrl` method to client (closes #11)
 * Fix error when calling `getTokens` on a card without multiple parts (closes #10)
@@ -6,7 +9,7 @@
 _Breaking Changes_
 * Remove `symbols` file in favor of `getSymbolUrl`
 
-# 1.11.0
+# 0.11.0
 * Normalize card_faces so all card objects have a card_faces array
 
 _Breaking Changes_